pyproject.toml
[build-system]
requires = ["hatchling", "tree_sitter_language_pack>=0.13.0"]
build-backend = "hatchling.build"

[project]
name = "gh-chat-dataset"
version = "0.1.0"
description = "One-shot tool to turn a GitHub repo (py/js + docs) into chat/messages JSONL for Qwen/MLX fine-tuning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Droid (Factory.ai)", email = "dev@factory.ai"}]
dependencies = [
  "click>=8.1.7",
  "anthropic>=0.31.0",
  "faiss-cpu>=1.8.0",
  "hdbscan>=0.8.38",
  "markdown-it-py>=3.0.0",
  "numpy>=1.26.0",
  "openai>=1.40.0",
  "scikit-learn>=1.5.0",
  "tree_sitter>=0.21.0",
  "tree_sitter_language_pack>=0.13.0",
  "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.6.9",
]
web = [
  "flask-cors>=4.0.0",
]

[project.scripts]
gh-chat-dataset = "gh_chat_dataset.cli:main"
gh-chat-dataset-web = "gh_chat_dataset.webapp.server:main"
calculator-app = "calculator_app.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gh_chat_dataset", "src/gh_chat_dataset/webapp", "src/calculator_app"]

[tool.hatch.build.targets.sdist]
include = [
  "src/gh_chat_dataset",
  "src/gh_chat_dataset/webapp",
  "src/calculator_app",
  "README.md",
  "pyproject.toml",
  "LICENSE",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
